Fix calculation to apply taxes after discount
Introduces the field "NetTotal" for orders and price items.
This field represents the net amount of a single item or an order
that taxes are calculated based on. While the "subtotal" amount is
not discounted, the "NetTotal" is.
This is especially important when prices are expected to include
taxes. In these cases discounts are applied to the original price
and then the net price is calculated.

The test values are synced with those in GoCommerce.